requirements.txt
streamlit
langchain
langchain_community
langchain-openai
openai
tiktoken
faiss-cpu
python-dotenv